Use mixins to avoid duplication.
